Moved BirthPreparednessFormsAggregationHelper to own file
