+++ custom/icds_reports/models/aggregate.py
@@ -13,7 +13,6 @@ from custom.icds_reports.const import (AGG_CCS_RECORD_BP_TABLE,
 from custom.icds_reports.utils.aggregation import (
     AggCcsRecordAggregationHelper, AggChildHealthAggregationHelper,
     AwcInfrastructureAggregationHelper,
-    BirthPreparednessFormsAggregationHelper,
     CcsRecordMonthlyAggregationHelper, ChildHealthMonthlyAggregationHelper,
     DailyFeedingFormsChildHealthAggregationHelper,
     DeliveryFormsAggregationHelper,
@@ -24,6 +23,8 @@ from dateutil.relativedelta import relativedelta
 from django.db import connections, models, transaction
 from six.moves import range
 
+from custom.icds_reports.utils.aggregation_helpers.birth_preparedness_forms import \
+    BirthPreparednessFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms import ComplementaryFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms_ccs_record import \
     ComplementaryFormsCcsRecordAggregationHelper
